Simplify flatten, remove flatMap.
Rather than allow customizable depth for flatten that defaults to +Inf, thisinstead only flattens a single level. That makes flatMap almost exactly map().flatten() with the exception of asserting Iterable and ignoring ConcatSpreadable, which is arguably more confusing than helpful.

Since this seeks to provide a core set of simple functions, flatMap is removed in favor of the easier to understand behavior of map().flatten().

Closes #4